+++ src/every.ts
@@ -1,5 +1,23 @@
+import isSync from "./_isSync.ts";
+
+export default every;
+
+/** @ignored */
+function every<T>(
+  iterable: Iterable<T>,
+  predicate: (value: T) => boolean,
+): boolean;
+
+/** @ignored */
+function every<T>(
+  iterable: AsyncIterable<T>,
+  predicate: (value: T) => boolean,
+): Promise<boolean>;
+
 /**
  * Returns `true` if `predicate(value)` returns true for every value in `iterable`, and false otherwise. Returns `true` for an empty iterable.
+ *
+ * Works with sync and async iterables. If passed an async iterable, returns a Promise for the result.
  *
  * @example
  * ```typescript
@@ -16,11 +34,30 @@
  * // => true
  * ```
  */
-export default function every<T>(
+function every<T>(
+  iterable: Iterable<T> | AsyncIterable<T>,
+  predicate: (value: T) => boolean,
+): boolean | Promise<boolean> {
+  return isSync(iterable)
+    ? everySync(iterable, predicate)
+    : everyAsync(iterable, predicate);
+}
+
+function everySync<T>(
   iterable: Iterable<T>,
   predicate: (value: T) => boolean,
 ): boolean {
   for (const value of iterable) {
+    if (!predicate(value)) return false;
+  }
+  return true;
+}
+
+async function everyAsync<T>(
+  iterable: AsyncIterable<T>,
+  predicate: (value: T) => boolean,
+): Promise<boolean> {
+  for await (const value of iterable) {
     if (!predicate(value)) return false;
   }
   return true;